Don't add limit to params array if no valid limit is set.
When an invalid limit is set with .limit(<0), then when the params
are built, a non-falsy (i.e., non-zero) limit is seen and is added
to the params array.

Added code to update the value based on the validity of the limit arg.
Added tests to cover this case.

Thanks for a great, clean query builder library. This is very helpful.
-david